+++ deepcave/config.py
@@ -57,6 +57,7 @@ class Config:
     FIGURE_MARGIN = dict(t=30, b=0, l=0, r=0)
     FIGURE_HEIGHT = "40vh"
     FIGURE_DOWNLOAD_SCALE = 4.0
+    FIGURE_FONT_SIZE = 20
 
     # Redis settings
     REDIS_PORT: int = 6379
@@ -133,10 +134,11 @@ class Config:
     @property
     def CONVERTERS(self) -> List[Type["Run"]]:
         """Get a list of available run converters."""
+        from deepcave.runs.converters.amltk import AMLTKRun
         from deepcave.runs.converters.bohb import BOHBRun
         from deepcave.runs.converters.dataframe import DataFrameRun
         from deepcave.runs.converters.deepcave import DeepCAVERun
         from deepcave.runs.converters.smac3v1 import SMAC3v1Run
         from deepcave.runs.converters.smac3v2 import SMAC3v2Run
 
-        return [DeepCAVERun, BOHBRun, SMAC3v1Run, SMAC3v2Run, DataFrameRun]
+        return [DeepCAVERun, BOHBRun, SMAC3v1Run, SMAC3v2Run, AMLTKRun, DataFrameRun]