+++ deepcave/runs/run.py
@@ -103,7 +103,7 @@ class Run(AbstractRun, ABC):
             serialized_objectives += [objective.to_json()]
 
         # Meta
-        self.meta = {"objectives": serialized_objectives, "budgets": []}
+        self.meta = {"objectives": serialized_objectives, "budgets": [], "seeds": []}
         self.meta.update(meta)
 
     @classmethod
@@ -199,6 +199,7 @@ class Run(AbstractRun, ABC):
         self,
         costs: Union[List[float], float],
         config: Union[Dict, Configuration],
+        seed: int,
         budget: float = np.inf,
         start_time: float = 0.0,
         end_time: float = 0.0,
@@ -210,7 +211,7 @@ class Run(AbstractRun, ABC):
         """
         Add a trial to the run.
 
-        If combination of config and budget already exists, it will be overwritten.
+        If combination of config, seed, and budget already exists, it will be overwritten.
         Not successful runs are added with `None` costs.
 
         Parameters
@@ -219,6 +220,10 @@ class Run(AbstractRun, ABC):
             Costs of the run. In case of multi-objective, a list of costs is expected.
         config : Union[Dict, Configuration]
             The corresponding configuration.
+        seed : int
+            Seed of the run.
+        budget : float, optional
+            Budget of the run. By default np.inf
         start_time : float, optional
             Start time. By default, 0.0
         end_time : float, optional
@@ -290,6 +295,7 @@ class Run(AbstractRun, ABC):
         trial = Trial(
             config_id=config_id,
             budget=budget,
+            seed=seed,
             costs=costs,
             start_time=np.round(start_time, 2),
             end_time=np.round(end_time, 2),
@@ -311,6 +317,11 @@ class Run(AbstractRun, ABC):
             self.meta["budgets"].sort()
 
         self._update_highest_budget(config_id, budget, status)
+
+        # Update seeds
+        if seed not in self.meta["seeds"]:
+            self.meta["seeds"].append(seed)
+            self.meta["seeds"].sort()
 
         # Update models
         # Problem: The model should not be in the cache.